src/main/python/requirements.txt
odfpy
zipfile36
chardet
pytest-shutil
